requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# AI and NLP
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10

openai>=1.6.1,<2.0.0
python-dotenv==1.0.0



# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dateutil==2.8.2
lxml==4.9.3
html5lib==1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

